Emit error signals when consuming MssqlResult
We now emit error signals when consuming MssqlResult instead of when consuming Statement.execute(). This commit introduces a behavioral change as we no longer use windowing and fail the main sequence but we emit a MssqlResult per parameter binding set from ParametrizedMssqlStatement. Failing a sequence with windowing always leads to onErrorDropped and a potential bubbling of the exception across the call. In our case, exceptions bubbled into StreamDecoder which caused a lockup during result processing.

[resolves #162]