Add support for statement timeout.
The driver now times out statements that didn't emit a response (time to first response) within the specified timeout. A statement that times out triggers an attention token to cancel the query on the server side.

All cancellation and discard operators are now attached in the …Flow classes to reduce duplications.

[resolves #213]

Signed-off-by: Mark Paluch <mpaluch@vmware.com>